+++ examples/3_simple_chatbot.py
@@ -1,3 +1,5 @@
+import os
+
 from llmbrix.agent import Agent
 from llmbrix.chat_history import ChatHistory
 from llmbrix.gpt_openai import GptOpenAI
@@ -7,7 +9,7 @@ SYSTEM_MSG = "Be super brief."
 HIST_LIMIT = 5
 MODEL = "gpt-4o-mini"
 
-gpt = GptOpenAI(model=MODEL)
+gpt = GptOpenAI.from_openai(model=MODEL, api_key=os.getenv("OPENAI_API_KEY"))
 chat_history = ChatHistory(max_turns=HIST_LIMIT)
 system_msg = SystemMsg(content=SYSTEM_MSG)
 agent = Agent(gpt=gpt, chat_history=chat_history, system_msg=system_msg)