+++ llmbrix/gpt_openai.py
@@ -1,7 +1,7 @@
 import json
 from typing import Optional, Type, TypeVar, cast
 
-from openai import AzureOpenAI, OpenAI
+from openai import AzureOpenAI, OpenAI, OpenAIError
 from openai.types.responses import ResponseFunctionToolCall
 from pydantic import BaseModel
 
@@ -11,65 +11,96 @@ from llmbrix.msg import AssistantMsg, Msg, ToolRequestMsg
 from llmbrix.tool import Tool
 
 T = TypeVar("T", bound=BaseModel)
-DEFAULT_TIMEOUT = 60
 
 
 class GptOpenAI:
     """
-    Wraps OpenAI GPT responses API.
+    GPT wrapper used to generate tokens via OpenAI API.
+    Supports non-reasoning GPT models, tool calling and structured outputs.
+    Internally uses responses API.
     """
 
     def __init__(
         self,
-        model: str = None,
+        model: str,
+        openai_client: OpenAI | AzureOpenAI = None,
         tools: list[Tool] = None,
         output_format: Optional[Type[T]] = None,
-        api_timeout: int = DEFAULT_TIMEOUT,
-        openai_client: OpenAI | AzureOpenAI = None,
+        api_timeout: int = 60,
         **responses_kwargs,
     ):
         """
         Parameters passed here will be set as defaults.
-
         Any value passed to these parameters in `generate()` will override these defaults.
 
-        Note `model` has to be set either here or in `generate()` function.
-
-        If no `openai_client` then `OPENAI_API_KEY` env variable must be set in order to
-        initialize default OpenAI client.
-
-        :param model: name of GPT model to use
-        :param tools: (optional) list of Tool child instances to register to LLM as tools to be used
+        :param model: name of GPT model to use. Reasoning models are not supported.
+        :param openai_client: `OpenAI` or `AzureOpenAI` object from official OpenAI SDK.
+                              If not provided then:
+                                  1. `OpenAI` client is initialized from `OPENAI_API_KEY` env var.
+                                  2. If `OPENAI_API_KEY` not provided either then` AzureOpenAI` client is
+                                     initialized from env vars. At minimum set following vars:
+                                        - `AZURE_OPENAI_API_KEY`
+                                        - `OPENAI_API_VERSION`
+                                        - `AZURE_OPENAI_ENDPOINT`
+                              Refer to official SDK `OpenAI` and `AzureOpenAI` class docs for more details.
+        :param tools: (optional) list of Tool instances to register to LLM as tools to be used
         :param output_format: (optional) Pydantic BaseModel instance to define output format of the LLM.
         :param api_timeout: timeout for OpenAI API in seconds. Default is 60s
-        :param openai_client: (optional) Specify custom OpenAI/AzureOpenAI client to be used.
-                              If none provided then default OpenAI() client is initialized.
-
-                              Example: Initialize AzureOpenAI client:
-
-                                ```python
-                                    import os
-                                    from openai import AzureOpenAI
-
-                                    client = AzureOpenAI(
-                                        api_key=os.getenv("AZURE_OPENAI_API_KEY"),
-                                        api_version="2024-07-01-preview",
-                                        azure_endpoint=os.getenv("AZURE_OPENAI_ENDPOINT")
-                                    )
-                                ```
         :param responses_kwargs: (optional) any additional kwargs to be passed to responses API.
-                                 Note if output format is defined responses.parse is used.
-
         """
         self.model = model
         self.tools = tools
         self.output_format = output_format
         self.api_timeout = api_timeout
         self.responses_kwargs = responses_kwargs
-        if openai_client is not None:
+        if openai_client:
             self.client = openai_client
         else:
-            self.client = OpenAI()
+            try:
+                self.client = OpenAI()
+            except OpenAIError:
+                self.client = AzureOpenAI()
+
+    @classmethod
+    def from_openai(cls, model: str, api_key: str, **kwargs) -> "GptOpenAI":
+        """
+        Initialize GptOpenAI instance with standard OpenAI API.
+
+        :param model: name of GPT model to use. Reasoning models are not supported.
+        :param api_key: OpenAI API key.
+        :param kwargs: Kwargs passed to __init__ method. See DocString of __init__ for details.
+
+        :return: Instance of GptOpenAI with standard OpenAI client.
+        """
+        client = OpenAI(api_key=api_key)
+        return cls(model=model, openai_client=client, **kwargs)
+
+    @classmethod
+    def from_azure_openai(
+        cls,
+        model: str,
+        api_key: str,
+        api_version: str,
+        azure_endpoint: str,
+        azure_deployment: str,
+        **kwargs,
+    ) -> "GptOpenAI":
+        """
+        Initialize GptOpenAI instance from standard Azure OpenAI API args.
+
+        :param model: name of GPT model to use. Reasoning models are not supported.
+        :param api_key: Azure OpenAI API key.
+        :param api_version: Azure OpenAI API version.
+        :param azure_endpoint: Azure OpenAI API endpoint.
+        :param azure_deployment: Azure OpenAI API deployment.
+        :param kwargs: Kwargs passed to __init__ method. See DocString of __init__ for details.
+
+        :return: Instance of GptOpenAI with AzureOpenAI client.
+        """
+        client = AzureOpenAI(
+            api_key=api_key, api_version=api_version, azure_endpoint=azure_endpoint, azure_deployment=azure_deployment
+        )
+        return cls(model=model, openai_client=client, **kwargs)
 
     def __call__(self, *args, **kwargs) -> GptResponse:
         """
@@ -82,7 +113,6 @@ class GptOpenAI:
     def generate(
         self,
         messages: list[Msg],
-        model: str = None,
         tools: list[Tool] = None,
         output_format: Optional[Type[T]] = None,
         api_timeout: int = None,
@@ -95,7 +125,6 @@ class GptOpenAI:
         If provided here, they override those defaults.
 
         :param messages: list of messages for LLM to be used as input.
-        :param model: name of GPT model to use
         :param tools: (optional) list of Tool child instances to register to LLM as tools to be used
         :param output_format: (optional) Pydantic BaseModel instance to define output format of the LLM.
         :param api_timeout: timeout for OpenAI API in seconds. Default is set to 60s.
@@ -109,28 +138,22 @@ class GptOpenAI:
 
         :raises OpenAIResponseError: if an error field is returned from OpenAI responses API call
         """
-        model = model or self.model
         tools = tools or self.tools
         output_format = output_format or self.output_format
         api_timeout = self.api_timeout if api_timeout is None else api_timeout
         responses_kwargs = {**self.responses_kwargs, **responses_kwargs}
-        if not model:
-            raise ValueError(
-                "No model name has been provided. Either set default model name in __init__() or pass "
-                "it into this function."
-            )
 
         messages = [m.to_openai() for m in messages]
         tool_schemas = [t.openai_schema for t in tools] if tools else []
 
         if output_format is None:
             response = self.client.responses.create(
-                input=messages, model=model, tools=tool_schemas, timeout=api_timeout, **responses_kwargs
+                input=messages, model=self.model, tools=tool_schemas, timeout=api_timeout, **responses_kwargs
             )
         else:
             response = self.client.responses.parse(
                 input=messages,
-                model=model,
+                model=self.model,
                 tools=tool_schemas,
                 text_format=output_format,
                 timeout=api_timeout,